minor advances
